+++ compiler/flags/IntelLLVM_Fortran.cmake
@@ -1,33 +1,65 @@
-if (CMAKE_Fortran_COMPILER_VERSION VERSION_LESS 15.1)
-  message(FATAL_ERROR "${CMAKE_Fortran_COMPILER_ID} version must be at least 15.1!")
+if (CMAKE_Fortran_COMPILER_VERSION VERSION_LESS 2025.1)
+  message(FATAL_ERROR "${CMAKE_Fortran_COMPILER_ID} version must be at least 2025.1!")
 endif()
 
+# ----------------------------------------------------------------------
+# Optimization levels
+# ----------------------------------------------------------------------
 set (FOPT0 "-O0")
 set (FOPT1 "-O1")
 set (FOPT2 "-O2")
 set (FOPT3 "-O3")
 set (FOPT4 "-O4")
+set (FFAST "-fast")
+
+# Debug info
 set (DEBINFO "-g")
 
+# ----------------------------------------------------------------------
+# Floating-point handling
+# ----------------------------------------------------------------------
 set (FPE0 "-fpe0")
 set (FPE1 "-fpe1")
 set (FPE3 "-fpe3")
-# ifx does not support -fp-model source (yet? see https://www.intel.com/content/www/us/en/develop/documentation/fortran-compiler-oneapi-dev-guide-and-reference/top/compiler-reference/compiler-options/floating-point-options/fp-model-fp.html)
-#set (FP_MODEL_SOURCE "-fp-model source")
-set (FP_MODEL_SOURCE "")
-set (FP_MODEL_STRICT "-fp-model strict")
-# ifx does not support -fp-model consistent (yet? see https://www.intel.com/content/www/us/en/develop/documentation/fortran-compiler-oneapi-dev-guide-and-reference/top/compiler-reference/compiler-options/floating-point-options/fp-model-fp.html)
-#set (FP_MODEL_CONSISTENT "-fp-model consistent")
-set (FP_MODEL_CONSISTENT "")
-set (FP_MODEL_FAST1 "-fp-model fast=1")
-set (FP_MODEL_FAST2 "-fp-model fast=2")
 
-set (OPTREPORT0 "-qopt-report0")
-set (OPTREPORT5 "-qopt-report5")
+# Grouped FP models
+set (FP_SOURCE     "-fp-model source")
+set (FP_CONSISTENT "-fp-model consistent")
+set (FP_EXCEPT     "-fp-model except")
+set (FP_PRECISE    "-fp-model precise")
+set (FP_STRICT     "-fp-model strict")
+set (FP_FAST       "-fp-model fast")
+set (FP_FAST1      "-fp-model fast=1")
+set (FP_FAST2      "-fp-model fast=2")
 
+# -fp-speculation=fast is the compiler default
+set (FP_SPECULATION_FAST   "-fp-speculation=fast")
+# Testing with ifx 2025.2 found these flags caused a lot
+# of ICEs. For now we turn off
+#set (FP_SPECULATION_SAFE "-fp-speculation=safe")
+#set (FP_SPECULATION_STRICT "-fp-speculation=strict")
+
+set (ARCH_CONSISTENCY "-fimf-arch-consistency=true")
+
+set (FTZ "-ftz")
+set (NO_PREC_DIV "-no-prec-div")
+set (USE_SVML "-fimf-use-svml=true ")
+
+set (IPO "-ipo")
+set (FMA "-fma")
+set (NO_FMA "-no-fma")
 set (FREAL8 "-r8")
 set (FINT8 "-i8")
 
+# ----------------------------------------------------------------------
+# Reports
+# ----------------------------------------------------------------------
+set (OPTREPORT0 "-qopt-report0")
+set (OPTREPORT5 "-qopt-report5")
+
+# ----------------------------------------------------------------------
+# Portability / source format
+# ----------------------------------------------------------------------
 set(PP "-fpp") # default for all other versions
 if(CMAKE_Fortran_COMPILER_VERSION VERSION_GREATER_EQUAL 2025.2 AND CMAKE_Fortran_COMPILER_VERSION VERSION_LESS 2025.3)
 
@@ -61,28 +93,35 @@ exec \"${cpp_exe}\" -P -traditional-cpp -undef \"$@\"
   endif()
 endif()
 
-set (MISMATCH "")
 set (BIG_ENDIAN "-convert big_endian")
 set (LITTLE_ENDIAN "-convert little_endian")
 set (EXTENDED_SOURCE "-extend-source")
 set (FIXED_SOURCE "-fixed")
+set (NO_RANGE_CHECK "")
+
+# ----------------------------------------------------------------------
+# Warnings / diagnostics
+# ----------------------------------------------------------------------
 set (DISABLE_FIELD_WIDTH_WARNING "-diag-disable 8291")
 set (DISABLE_GLOBAL_NAME_WARNING "-diag-disable 5462")
-set (CRAY_POINTER "")
-set (MCMODEL "-mcmodel medium -shared-intel")
-set (HEAPARRAYS "-heap-arrays 32")
-set (BYTERECLEN "-assume byterecl")
-set (ALIGNCOM "-align dcommons")
-set (TRACEBACK "-traceback")
-set (NOOLD_MAXMINLOC "-assume noold_maxminloc")
-set (REALLOC_LHS "-assume realloc_lhs")
-set (ARCH_CONSISTENCY "-fimf-arch-consistency=true")
-set (FTZ "-ftz")
-set (ALIGN_ALL "-align all")
-set (NO_ALIAS "-fno-alias")
-set (USE_SVML "-fimf-use-svml=true")
+set (DISABLE_10337 "-diag-disable 10337")   # fno-builtin warning
+set (DISABLE_10121 "-diag-disable 10121")   # fp-model override warning
+set (DISABLE_10448 "-diag-disable 10448")   # ifort deprecation remark
+set (DISABLE_LONG_LINE_LENGTH_WARNING "-diag-disable 5268")
 
-# Additional flags for better Standards compliance
+# Make an option to make things quiet during debug builds
+option (QUIET_DEBUG "Suppress excess compiler output during debug builds" OFF)
+if (QUIET_DEBUG)
+  set (WARN_UNUSED "")
+  set (SUPPRESS_COMMON_WARNINGS "${DISABLE_FIELD_WIDTH_WARNING} ${DISABLE_GLOBAL_NAME_WARNING} ${DISABLE_10337}")
+else ()
+  set (WARN_UNUSED "-warn unused")
+  set (SUPPRESS_COMMON_WARNINGS "${DISABLE_GLOBAL_NAME_WARNING} ${DISABLE_10337}")
+endif ()
+
+# ----------------------------------------------------------------------
+# Standards Compliance (used in MAPL)
+# ----------------------------------------------------------------------
 ## Set the Standard to be Fortran 2018
 set (STANDARD_F18 "-stand f18")
 ## Error out if you try to do if(integer)
@@ -92,7 +131,20 @@ set (ERROR_LOGICAL_SET_TO_INTEGER "-diag-error 6192")
 ## Turn off warning #5268 (Extension to standard: The text exceeds right hand column allowed on the line.)
 set (DISABLE_LONG_LINE_LENGTH_WARNING "-diag-disable 5268")
 
-set (NO_RANGE_CHECK "")
+# ----------------------------------------------------------------------
+# Memory / alignment
+# ----------------------------------------------------------------------
+set (MCMODEL "-mcmodel medium -shared-intel")
+set (HEAPARRAYS "-heap-arrays 32")
+set (BYTERECLEN "-assume byterecl")
+set (TRACEBACK "-traceback")
+set (NOOLD_MAXMINLOC "-assume noold_maxminloc")
+set (REALLOC_LHS "-assume realloc_lhs")
+set (ALIGNCOM "-align dcommons")
+set (NO_ALIAS "-fno-alias")
+set (ALIGN_ALL "-align all")
+set (ARRAY_ALIGN_32BYTE "-align array32byte")
+set (ARRAY_ALIGN_64BYTE "-align array64byte")
 
 cmake_host_system_information(RESULT proc_description QUERY PROCESSOR_DESCRIPTION)
 if (${proc_description} MATCHES "EPYC")
@@ -112,43 +164,62 @@ else ()
   message(FATAL_ERROR "Unknown processor. Please file an issue at https://github.com/GEOS-ESM/ESMA_cmake")
 endif ()
 
+# ----------------------------------------------------------------------
+# Defines
+# ----------------------------------------------------------------------
 add_definitions(-DHAVE_SHMEM)
 
-####################################################
+# ----------------------------------------------------------------------
+# Common flag bundles
+# ----------------------------------------------------------------------
+set (common_Fortran_flags "${TRACEBACK} ${REALLOC_LHS} ${OPTREPORT0} ${ALIGN_ALL} ${NO_ALIAS} ${PP}")
+set (common_Fortran_fpe_flags "${FTZ} ${NOOLD_MAXMINLOC}")
 
-# Common Fortran Flags
-# --------------------
-set (common_Fortran_flags "${TRACEBACK} ${REALLOC_LHS} ${PP}")
-set (common_Fortran_fpe_flags "${FPE0} ${FP_MODEL_SOURCE} ${HEAPARRAYS} ${NOOLD_MAXMINLOC}")
+# ----------------------------------------------------------------------
+# Build type specific bundles
+# ----------------------------------------------------------------------
+# Debug
+set (GEOS_Fortran_Debug_Flags "${DEBINFO} ${FOPT0} -debug -nolib-inline -fno-inline-functions -assume protect_parens,minus0 -prec-div -check all,noarg_temp_created,nouninit ${WARN_UNUSED} -init=snan,arrays -save-temps")
+set (GEOS_Fortran_Debug_FPE_Flags "${FPE0} ${FP_STRICT} ${FP_SPECULATION_STRICT} ${common_Fortran_fpe_flags} ${SUPPRESS_COMMON_WARNINGS}")
 
-# GEOS Debug
-# ----------
-set (GEOS_Fortran_Debug_Flags "${DEBINFO} ${FOPT0} ${FTZ} ${ALIGN_ALL} ${NO_ALIAS} -debug -nolib-inline -fno-inline-functions -assume protect_parens,minus0 -prec-div -check all,noarg_temp_created,nouninit -warn unused -save-temps")
-set (GEOS_Fortran_Debug_FPE_Flags "${common_Fortran_fpe_flags}")
+# GEOS Safe
+# ----------------
+set (GEOS_Fortran_Safe_Flags "${FOPT2} ${DEBINFO}")
+set (GEOS_Fortran_Safe_FPE_Flags "${FPE1} ${FP_PRECISE} ${FP_SOURCE} ${FP_CONSISTENT} ${NO_FMA} ${ARCH_CONSISTENCY} ${FP_SPECULATION_STRICT} ${common_Fortran_fpe_flags}")
 
 # GEOS NoVectorize
 # ----------------
-set (GEOS_Fortran_NoVect_Flags "${FOPT3} ${DEBINFO} ${OPTREPORT0} ${FTZ} ${ALIGN_ALL} ${NO_ALIAS}")
-set (GEOS_Fortran_NoVect_FPE_Flags "${common_Fortran_fpe_flags} ${ARCH_CONSISTENCY}")
+set (GEOS_Fortran_NoVect_Flags "${FOPT3} ${DEBINFO}")
+set (GEOS_Fortran_NoVect_FPE_Flags "${FPE1} ${FP_FAST1} ${FP_SOURCE} ${FP_CONSISTENT} ${NO_FMA} ${ARCH_CONSISTENCY} ${FP_SPECULATION_SAFE} ${common_Fortran_fpe_flags}")
 
 # NOTE It was found that the Vectorizing Flags gave better performance with the same results in testing.
 #      But in case they are needed, we keep the older flags available
 
-# GEOS Vectorize
-# --------------
-set (GEOS_Fortran_Vect_Flags "${FOPT3} ${DEBINFO} ${MARCH_FLAG} -fma -qopt-report0 ${FTZ} ${ALIGN_ALL} ${NO_ALIAS} -align array32byte")
-set (GEOS_Fortran_Vect_FPE_Flags "${FPE3} ${FP_MODEL_CONSISTENT} ${NOOLD_MAXMINLOC}")
+# GEOS Stock-Vect
+# ---------------
+set (GEOS_Fortran_StockVect_Flags "${FOPT3} ${DEBINFO} ${MARCH_FLAG} ${FMA} -align array32byte")
+set (GEOS_Fortran_StockVect_FPE_Flags "${FPE3} ${FP_FAST} ${FP_SOURCE} ${FP_CONSISTENT} ${common_Fortran_fpe_flags}")
 
-# GEOS Release
-# ------------
-set (GEOS_Fortran_Release_Flags  "${GEOS_Fortran_Vect_Flags}")
-set (GEOS_Fortran_Release_FPE_Flags "${GEOS_Fortran_Vect_FPE_Flags}")
+# GEOS Vectorize
+# ---------------
+set (GEOS_Fortran_Vect_Flags "${FOPT3} ${DEBINFO} ${MARCH_FLAG} -align array32byte")
+set (GEOS_Fortran_Vect_FPE_Flags "${FPE1} ${FP_FAST1} ${FP_SOURCE} ${FP_CONSISTENT} ${NO_FMA} ${ARCH_CONSISTENCY} ${FP_SPECULATION_SAFE} ${common_Fortran_fpe_flags}")
+
+# GEOS VectTrap
+# -------------
+# Until these can be tested, we just use the same as GEOS_Vectorize
+set (GEOS_Fortran_VectTrap_Flags "${FOPT3} ${DEBINFO} ${MARCH_FLAG} -align array32byte")
+set (GEOS_Fortran_VectTrap_FPE_Flags "${FPE1} ${FP_FAST1} ${FP_SOURCE} ${FP_CONSISTENT} ${NO_FMA} ${ARCH_CONSISTENCY} ${FP_SPECULATION_SAFE} ${common_Fortran_fpe_flags}")
 
 # GEOS Aggressive
 # ---------------
-set (GEOS_Fortran_Aggressive_Flags "${FOPT3} ${DEBINFO} ${MARCH_FLAG} -fma -qopt-report0 ${FTZ} ${ALIGN_ALL} ${NO_ALIAS} -align array32byte")
-#set (GEOS_Fortran_Aggressive_Flags "${FOPT3} ${DEBINFO} -xSKYLAKE-AVX512 -qopt-zmm-usage=high -fma -qopt-report0 ${FTZ} ${ALIGN_ALL} ${NO_ALIAS} -align array64byte")
-set (GEOS_Fortran_Aggressive_FPE_Flags "${FPE3} ${FP_MODEL_FAST2} ${USE_SVML} ${NOOLD_MAXMINLOC}")
+set (GEOS_Fortran_Aggressive_Flags "${FOPT3} ${DEBINFO} ${MARCH_FLAG} -align array32byte")
+set (GEOS_Fortran_Aggressive_FPE_Flags "${FPE3} ${FP_FAST2} ${FP_SOURCE} ${FP_CONSISTENT} ${FMA} ${FP_SPECULATION_FAST} ${USE_SVML} ${common_Fortran_fpe_flags}")
+
+# Set Release flags
+# -----------------
+set (GEOS_Fortran_Release_Flags  "${GEOS_Fortran_Vect_Flags}")
+set (GEOS_Fortran_Release_FPE_Flags "${GEOS_Fortran_Vect_FPE_Flags}")
 
 # Common variables for every compiler
 include(Generic_Fortran)